arm64: prefetch: Change assembly to be compatible with gcc and clang
clang fails to build with the current code:

arch/arm64/include/asm/processor.h:172:15: error: invalid operand in
inline asm: 'prfm pldl1keep, ${0:a}'

Apparently clang does not support the 'a' modifier. Change the
constraint from 'p' ('An operand that is a valid memory address is
allowed') to 'Q' ('A memory address which uses a single base register
with no offset'), which works for both gcc and clang.

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
https://patchwork.kernel.org/project/linux-arm-kernel/patch/20170419212211.95803-1-mka@chromium.org/
Signed-off-by: crawford. <sh1no.crawford@gmail.com>